+++ multimux/tests/golden_gate.rs
@@ -46,6 +46,7 @@ use axum::response::IntoResponse;
 use axum::routing::get;
 use serde_json::Value;
 
+use broadcast_hls::{MapTag, MediaPlaylist, MediaSegment};
 use hls_runtime::client::Output;
 use hls_runtime::client::tokio_client::TokioClient;
 use hls_runtime::server::DEFAULT_TRACK_ID;
@@ -53,7 +54,6 @@ use multimux::origin::{AppState, router};
 use multimux::output::Output as MmOutput;
 use multimux::output::llhls::LlHlsOutput;
 use multimux::route::RouteHandle;
-use transmux::hls::{MapTag, MediaPlaylist, MediaSegment};
 use transmux::ll_hls::LlHlsSegmenter;
 use transmux::{CodecConfig, FragmentTrackData, Sample, TrackSpec, TsDemux};